improvement: fall back to AngularJS module name suggestion macro in power save mode
module name suggestion from AngularTemplates is a little more efficient in a sense that it suggests more relevant and nearest modules first. but it needs a little bit of calculation which is not optimized yet. So we would better to turn it of in favour of AngularJS module name autocompletion contributor if powerSaveMode is enabled.